Move ScoreAlgorithms class to an independent source file
