+++ src/graphic_model/lomse_selections.cpp
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 // This file is part of the Lomse library.
-// Lomse is copyrighted work (c) 2010-2016. All rights reserved.
+// Lomse is copyrighted work (c) 2010-2018. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -35,6 +35,7 @@
 #include "lomse_staffobjs_table.h"
 #include "lomse_document.h"
 #include "lomse_graphical_model.h"
+#include "lomse_score_algorithms.h"
 
 #include <sstream>
 using namespace std;
