+++ src/tests/lomse_test_score_algorithms.cpp
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 // This file is part of the Lomse library.
-// Lomse is copyrighted work (c) 2010-2016. All rights reserved.
+// Lomse is copyrighted work (c) 2010-2018. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -38,6 +38,7 @@
 #include "lomse_im_note.h"
 #include "lomse_staffobjs_table.h"
 #include "lomse_document_cursor.h"
+#include "lomse_score_algorithms.h"
 
 
 using namespace UnitTest;
